Reference (API)

API documentation for Precompiler.

Precompiler.@cache

Precompiler.@setup

Precompiler.@cache — Macro

Precompiler.@cache f(args...)

precompile (and save in the cache file) any method-calls that occur inside the expression. All calls (direct or indirect) inside a Precompiler.@cache block will be cached.

Precompiler.@cache has three key features:

code inside runs only when the package is being precompiled (i.e., a *.ji precompile cache file is being written)

the interpreter is disabled, ensuring your calls will be compiled

both direct and indirect callees will be precompiled, even for methods defined in other packages and even for runtime-dispatched callees (requires Julia 1.8 and above).

Note

For comprehensive precompilation, ensure the first usage of a given method/argument-type combination occurs inside Precompiler.@cache.

In detail: runtime-dispatched callees are captured only when type-inference is executed, and they are inferred only on first usage. Inferrable calls that trace back to a method defined in your package, and their inferrable callees, will be precompiled regardless of "ownership" of the callees (Julia 1.8 and higher).

Consequently, this recommendation matters only for:

- direct calls to methods defined in Base or other packages OR
- indirect runtime-dispatched calls to such methods.

Precompiler.@setup — Macro

Precompiler.@setup begin
    vars = ...
    ⋮
end

Run the code block only during package precompilation. Precompiler.@setup is often used in combination with Precompiler.@cache, for example:

Precompiler.@setup begin
    vars = ...
    @cache begin
        y = f(vars...)
        g(y)
        ⋮
    end
end

Precompiler.@setup does not force compilation (though it may happen anyway) nor intentionally capture runtime dispatches (though they will be precompiled anyway if the runtime-callee is for a method belonging to your package).